Changed: Extract Class `TextFinder` from `TerminalBuffer`
Since do too much work in one big class, extracted the method to find the selected text and made it into a class.
